added online getting tickets

+++ data/src/main/java/com/example/data/repositories/UserRepository.kt
@@ -15,7 +15,7 @@ class UserRepository @Inject constructor(
     override suspend fun getTestUsers(): List<User> {
         return List(15) {
             User(
-                id = it,
+                id = it.toLong(),
                 firstName = "Ikhsanov$it",
                 name = "Ruslan$it",
                 lastName = "Lenarovich$it"
